Partially working slicing and data access.
