+++ scippy/variable.cpp
@@ -11,8 +11,11 @@
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
 
+#include "bind_data_access.h"
+#include "bind_slice_methods.h"
 #include "dataset.h"
 #include "except.h"
+#include "numpy.h"
 #include "scipp/units/unit.h"
 #include "tag_util.h"
 #include "variable.h"
@@ -22,63 +25,6 @@ using namespace scipp::core;
 using namespace scipp::units;
 
 namespace py = pybind11;
-
-template <typename T, typename SZ_TP>
-Variable makeVariableFromBuffer(const Dimensions &dimensions,
-                                const std::vector<SZ_TP> &stridesInBytes,
-                                T *ptr) {
-  auto ndims = scipp::size(dimensions.shape());
-  if (ndims == 0) // empty dataset
-    return makeVariable<underlying_type_t<T>>(dimensions);
-
-  std::vector<SZ_TP> varStrides(ndims, 1), strides;
-  for (auto &&strd : stridesInBytes)
-    strides.emplace_back(strd / sizeof(T));
-
-  bool sameStrides{*strides.rbegin() == 1};
-  auto i = varStrides.size() - 1;
-  while (i-- > 0) {
-    varStrides[i] = varStrides[i + 1] * dimensions.size(i + 1);
-    if (varStrides[i] != strides[i] && sameStrides)
-      sameStrides = false;
-  }
-
-  if (sameStrides) { // memory is alligned c-style and dense
-    return Variable(
-        units::dimensionless, std::move(dimensions),
-        Vector<underlying_type_t<T>>(ptr, ptr + dimensions.volume()));
-
-  } else {
-    // Try to find blocks to copy
-    auto index = scipp::size(strides) - 1;
-    while (strides[index] == varStrides[index])
-      --index;
-    ++index;
-    auto blockSz = index < scipp::size(strides)
-                       ? strides[index] * dimensions.size(index)
-                       : 1;
-
-    auto res = makeVariable<underlying_type_t<T>>(dimensions);
-    std::vector<scipp::index> dsz(ndims);
-    for (scipp::index i = 0; i < index; ++i)
-      dsz[i] = dimensions.size(i);
-    std::vector<scipp::index> coords(ndims, 0);
-    auto nBlocks = dimensions.volume() / blockSz;
-
-    for (scipp::index i = 0; i < nBlocks; ++i) {
-      // calculate the array linear coordinate
-      auto lin_coord = std::inner_product(coords.begin(), coords.end(),
-                                          strides.begin(), scipp::index{0});
-      std::memcpy(&res.template values<T>()[i * blockSz], &ptr[lin_coord],
-                  blockSz * sizeof(T));
-      // get the next ND coordinate
-      auto k = coords.size();
-      while (k-- > 0)
-        ++coords[k] >= dsz[k] ? coords[k] = 0 : k = 0;
-    }
-    return res;
-  }
-}
 
 template <typename Collection>
 auto getItemBySingleIndex(Collection &self,
@@ -93,47 +39,6 @@ auto getItemBySingleIndex(Collection &self,
   if (idx < 0)
     idx = sz + idx;
   return self(std::get<Dim>(index), idx);
-}
-
-template <class T> struct mutable_span_methods {
-  static void add(py::class_<scipp::span<T>> &span) {
-    span.def("__setitem__", [](scipp::span<T> &self, const scipp::index i,
-                               const T value) { self[i] = value; });
-  }
-};
-template <class T> struct mutable_span_methods<const T> {
-  static void add(py::class_<scipp::span<const T>> &) {}
-};
-
-template <class T> void declare_span(py::module &m, const std::string &suffix) {
-  py::class_<scipp::span<T>> span(m, (std::string("span_") + suffix).c_str());
-  span.def("__getitem__", &scipp::span<T>::operator[],
-           py::return_value_policy::reference)
-      .def("size", &scipp::span<T>::size)
-      .def("__len__", &scipp::span<T>::size)
-      .def("__iter__",
-           [](const scipp::span<T> &self) {
-             return py::make_iterator(self.begin(), self.end());
-           })
-      .def("__repr__",
-           [](const scipp::span<T> &self) { return array_to_string(self); });
-  mutable_span_methods<T>::add(span);
-}
-
-template <class T>
-void declare_VariableView(py::module &m, const std::string &suffix) {
-  py::class_<VariableView<T>> view(
-      m, (std::string("VariableView_") + suffix).c_str());
-  view.def("__repr__",
-           [](const VariableView<T> &self) { return array_to_string(self); })
-      .def("__getitem__", &VariableView<T>::operator[],
-           py::return_value_policy::reference)
-      .def("__setitem__", [](VariableView<T> &self, const scipp::index i,
-                             const T value) { self[i] = value; })
-      .def("__len__", &VariableView<T>::size)
-      .def("__iter__", [](const VariableView<T> &self) {
-        return py::make_iterator(self.begin(), self.end());
-      });
 }
 
 template <class T> struct MakeVariable {
@@ -229,26 +134,6 @@ std::vector<scipp::index> numpy_strides(const std::vector<scipp::index> &s) {
   return strides;
 }
 
-template <class T> struct SetData {
-  static void apply(const VariableProxy &slice, const py::array &data) {
-    // Pybind11 converts py::array to py::array_t for us, with all sorts of
-    // automatic conversions such as integer to double, if required.
-    py::array_t<T> dataT(data);
-
-    const auto &dims = slice.dims();
-    const py::buffer_info info = dataT.request();
-    const auto &shape = dims.shape();
-    if (!std::equal(info.shape.begin(), info.shape.end(), shape.begin(),
-                    shape.end()))
-      throw std::runtime_error(
-          "Shape mismatch when setting data from numpy array.");
-
-    auto buf = slice.values<T>();
-    auto *ptr = (T *)info.ptr;
-    std::copy(ptr, ptr + slice.dims().volume(), buf.begin());
-  }
-};
-
 template <class T>
 auto pySlice(T &source, const std::tuple<Dim, const py::slice> &index)
     -> decltype(source(Dim::Invalid, 0)) {
@@ -260,22 +145,6 @@ auto pySlice(T &source, const std::tuple<Dim, const py::slice> &index)
   if (step != 1)
     throw std::runtime_error("Step must be 1");
   return source(dim, start, stop);
-}
-
-void setVariableProxy(VariableProxy &self,
-                      const std::tuple<Dim, scipp::index> &index,
-                      const py::array &data) {
-  auto slice = self(std::get<Dim>(index), std::get<scipp::index>(index));
-  CallDType<double, float, int64_t, int32_t, bool>::apply<SetData>(
-      slice.dtype(), slice, data);
-}
-
-void setVariableProxyRange(VariableProxy &self,
-                           const std::tuple<Dim, const py::slice> &index,
-                           const py::array &data) {
-  auto slice = pySlice(self, index);
-  CallDType<double, float, int64_t, int32_t, bool>::apply<SetData>(
-      slice.dtype(), slice, data);
 }
 
 template <class T> struct MakePyBufferInfoT {
@@ -327,115 +196,6 @@ std::variant<py::array_t<Ts>...> as_py_array_t_variant(py::object &obj) {
   }
 }
 
-template <class... Ts> struct as_VariableViewImpl {
-  template <class Var>
-  static std::variant<std::conditional_t<
-      std::is_same_v<Var, Variable>, scipp::span<underlying_type_t<Ts>>,
-      VariableView<underlying_type_t<Ts>>>...>
-  values(Var &view) {
-    switch (view.dtype()) {
-    case dtype<double>:
-      return {view.template values<double>()};
-    case dtype<float>:
-      return {view.template values<float>()};
-    case dtype<int64_t>:
-      return {view.template values<int64_t>()};
-    case dtype<int32_t>:
-      return {view.template values<int32_t>()};
-    case dtype<bool>:
-      return {view.template values<bool>()};
-    case dtype<std::string>:
-      return {view.template values<std::string>()};
-    case dtype<boost::container::small_vector<double, 8>>:
-      return {
-          view.template values<boost::container::small_vector<double, 8>>()};
-    case dtype<Dataset>:
-      return {view.template values<Dataset>()};
-    case dtype<Eigen::Vector3d>:
-      return {view.template values<Eigen::Vector3d>()};
-    default:
-      throw std::runtime_error("not implemented for this type.");
-    }
-  }
-
-  template <class Var>
-  static std::variant<std::conditional_t<
-      std::is_same_v<Var, Variable>, scipp::span<underlying_type_t<Ts>>,
-      VariableView<underlying_type_t<Ts>>>...>
-  variances(Var &view) {
-    switch (view.dtype()) {
-    case dtype<double>:
-      return {view.template values<double>()};
-    case dtype<float>:
-      return {view.template values<float>()};
-    case dtype<int64_t>:
-      return {view.template values<int64_t>()};
-    case dtype<int32_t>:
-      return {view.template values<int32_t>()};
-    case dtype<bool>:
-      return {view.template values<bool>()};
-    case dtype<std::string>:
-      return {view.template values<std::string>()};
-    case dtype<boost::container::small_vector<double, 8>>:
-      return {
-          view.template values<boost::container::small_vector<double, 8>>()};
-    case dtype<Dataset>:
-      return {view.template values<Dataset>()};
-    case dtype<Eigen::Vector3d>:
-      return {view.template values<Eigen::Vector3d>()};
-    default:
-      throw std::runtime_error("not implemented for this type.");
-    }
-  }
-
-  // Return a scalar value from a variable, implicitly requiring that the
-  // variable is 0-dimensional and thus has only a single item.
-  template <class Var> static py::object value(Var &view) {
-    expect::equals(Dimensions(), view.dims());
-    return std::visit(
-        [](const auto &data) {
-          return py::cast(data[0], py::return_value_policy::reference);
-        },
-        values(view));
-  }
-  // Return a scalar variance from a variable, implicitly requiring that the
-  // variable is 0-dimensional and thus has only a single item.
-  template <class Var> static py::object variance(Var &view) {
-    expect::equals(Dimensions(), view.dims());
-    return std::visit(
-        [](const auto &data) {
-          return py::cast(data[0], py::return_value_policy::reference);
-        },
-        variances(view));
-  }
-  // Set a scalar value in a variable, implicitly requiring that the
-  // variable is 0-dimensional and thus has only a single item.
-  template <class Var> static void set_value(Var &view, const py::object &o) {
-    expect::equals(Dimensions(), view.dims());
-    std::visit(
-        [&o](const auto &data) {
-          data[0] = o.cast<typename std::decay_t<decltype(data)>::value_type>();
-        },
-        values(view));
-  }
-  // Set a scalar variance in a variable, implicitly requiring that the
-  // variable is 0-dimensional and thus has only a single item.
-  template <class Var>
-  static void set_variance(Var &view, const py::object &o) {
-    expect::equals(Dimensions(), view.dims());
-    std::visit(
-        [&o](const auto &data) {
-          data[0] = o.cast<typename std::decay_t<decltype(data)>::value_type>();
-        },
-        variances(view));
-  }
-};
-
-using as_VariableView =
-    as_VariableViewImpl<double, float, int64_t, int32_t, bool, std::string,
-                        boost::container::small_vector<double, 8>, Dataset,
-                        Eigen::Vector3d>;
-
 using small_vector = boost::container::small_vector<double, 8>;
 PYBIND11_MAKE_OPAQUE(small_vector);
 
@@ -443,30 +203,8 @@ void init_variable(py::module &m) {
   py::bind_vector<boost::container::small_vector<double, 8>>(
       m, "SmallVectorDouble8");
 
-  declare_span<double>(m, "double");
-  declare_span<float>(m, "float");
-  declare_span<Bool>(m, "bool");
-  declare_span<const double>(m, "double_const");
-  declare_span<const long>(m, "long_const");
-  declare_span<long>(m, "long");
-  declare_span<const std::string>(m, "string_const");
-  declare_span<std::string>(m, "string");
-  declare_span<const Dim>(m, "Dim_const");
-  declare_span<Dataset>(m, "Dataset");
-  declare_span<Eigen::Vector3d>(m, "Eigen_Vector3d");
-
-  declare_VariableView<double>(m, "double");
-  declare_VariableView<float>(m, "float");
-  declare_VariableView<int64_t>(m, "int64");
-  declare_VariableView<int32_t>(m, "int32");
-  declare_VariableView<std::string>(m, "string");
-  declare_VariableView<Bool>(m, "bool");
-  declare_VariableView<boost::container::small_vector<double, 8>>(
-      m, "SmallVectorDouble8");
-  declare_VariableView<Dataset>(m, "Dataset");
-  declare_VariableView<Eigen::Vector3d>(m, "Eigen_Vector3d");
-
-  py::class_<Variable>(m, "Variable")
+  py::class_<Variable> variable(m, "Variable");
+  variable
       .def(py::init(&makeVariableDefaultInit),
            py::arg("labels") = std::vector<Dim>{},
            py::arg("shape") = std::vector<scipp::index>{},
@@ -521,20 +259,6 @@ void init_variable(py::module &m) {
           &as_py_array_t_variant<Variable, double, float, int64_t, int32_t,
                                  bool>,
           "Returns a read-only numpy array containing the Variable's values.")
-      .def_property_readonly(
-          "values", &as_VariableView::values<Variable>,
-          "Returns a read-only VariableView onto the Variable's contents.")
-      .def_property_readonly(
-          "variances", &as_VariableView::variances<Variable>,
-          "Returns a read-only VariableView onto the Variable's contents.")
-      .def_property("value", &as_VariableView::value<Variable>,
-                    &as_VariableView::set_value<Variable>,
-                    "The only value for a 0-dimensional variable. Raises an "
-                    "exception if the variable is not 0-dimensional.")
-      .def_property("variances", &as_VariableView::variances<Variable>,
-                    &as_VariableView::set_variance<Variable>,
-                    "The only data point for a 0-dimensional variable. Raises "
-                    "an exception if the variable is not 0-dimensional.")
       .def(py::self == py::self, py::call_guard<py::gil_scoped_release>())
       .def(py::self + py::self, py::call_guard<py::gil_scoped_release>())
       .def(py::self + double(), py::call_guard<py::gil_scoped_release>())
@@ -583,9 +307,11 @@ void init_variable(py::module &m) {
       .def("__repr__",
            [](const Variable &self) { return to_string(self, "."); });
 
-  py::class_<VariableProxy> view(m, "VariableProxy", py::buffer_protocol());
-  view.def_buffer(&make_py_buffer_info);
-  view.def_property_readonly(
+  py::class_<VariableProxy> variableProxy(m, "VariableProxy",
+                                          py::buffer_protocol());
+  variableProxy.def_buffer(&make_py_buffer_info);
+  variableProxy
+      .def_property_readonly(
           "dims", [](const VariableProxy &self) { return self.dims(); },
           py::return_value_policy::copy,
           "A read-only Dimensions object containing the dimensions of the "
@@ -618,8 +344,6 @@ void init_variable(py::module &m) {
              const auto & [ dim, i ] = index;
              self(dim, i).assign(other);
            })
-      .def("__setitem__", &setVariableProxy)
-      .def("__setitem__", &setVariableProxyRange)
       .def("copy", [](const VariableProxy &self) { return Variable(self); },
            "Make a copy of a VariableProxy and return it as a Variable.")
       .def("__copy__", [](VariableProxy &self) { return Variable(self); })
@@ -631,22 +355,6 @@ void init_variable(py::module &m) {
                                  bool>,
           "Returns a read-only numpy array containing the VariableProxy's "
           "values.")
-      .def_property_readonly(
-          "values", &as_VariableView::values<VariableProxy>,
-          "Returns a read-only VariableView onto the VariableProxy's contents.")
-      .def_property_readonly(
-          "variances", &as_VariableView::variances<VariableProxy>,
-          "Returns a read-only VariableView onto the VariableProxy's contents.")
-      .def_property("value", &as_VariableView::value<VariableProxy>,
-                    &as_VariableView::set_value<VariableProxy>,
-                    "The only data point for a 0-dimensional "
-                    "variable. Raises an exception of the variable is "
-                    "not 0-dimensional.")
-      .def_property("variances", &as_VariableView::variances<VariableProxy>,
-                    &as_VariableView::set_variance<VariableProxy>,
-                    "The only data point for a 0-dimensional "
-                    "variable. Raises an exception of the variable is "
-                    "not 0-dimensional.")
       .def(py::self += py::self, py::call_guard<py::gil_scoped_release>())
       .def(py::self -= py::self, py::call_guard<py::gil_scoped_release>())
       .def(py::self *= py::self, py::call_guard<py::gil_scoped_release>())
@@ -699,6 +407,9 @@ void init_variable(py::module &m) {
            })
       .def("__repr__",
            [](const VariableProxy &self) { return to_string(self, "."); });
+
+  bind_data_properties(variable);
+  bind_data_properties(variableProxy);
 
   // Implicit conversion VariableProxy -> Variable. Reduces need for excessive
   // operator overload definitions
